feat: slash/** => slash
Move the commands into the root command directory, clean up
loading helpers.
