+++ src/commands/code.ts
@@ -2,12 +2,12 @@ import {
   SlashCommandBuilder,
   SlashCommandSubcommandBuilder,
 } from "@discordjs/builders";
-import { SlashInt } from "../../interfaces/slash/SlashInt";
-import { errorEmbedGenerator } from "../../modules/commands/errorEmbedGenerator";
-import { handleCanIUse } from "../../modules/commands/subcommands/code/handleCanIUse";
-import { handleColour } from "../../modules/commands/subcommands/code/handleColour";
-import { handleHttp } from "../../modules/commands/subcommands/code/handleHttp";
-import { beccaErrorHandler } from "../../utils/beccaErrorHandler";
+import { SlashInt } from "../interfaces/slash/SlashInt";
+import { errorEmbedGenerator } from "../modules/commands/errorEmbedGenerator";
+import { handleCanIUse } from "../modules/commands/subcommands/code/handleCanIUse";
+import { handleColour } from "../modules/commands/subcommands/code/handleColour";
+import { handleHttp } from "../modules/commands/subcommands/code/handleHttp";
+import { beccaErrorHandler } from "../utils/beccaErrorHandler";
 
 export const code: SlashInt = {
   data: new SlashCommandBuilder()
