+++ src/commands/manage.ts
@@ -2,12 +2,12 @@ import {
   SlashCommandBuilder,
   SlashCommandSubcommandBuilder,
 } from "@discordjs/builders";
-import { SlashInt } from "../../interfaces/slash/SlashInt";
-import { errorEmbedGenerator } from "../../modules/commands/errorEmbedGenerator";
-import { handleResetLevels } from "../../modules/commands/subcommands/manage/handleResetLevels";
-import { handleResetStars } from "../../modules/commands/subcommands/manage/handleResetStars";
-import { handleSuggestion } from "../../modules/commands/subcommands/manage/handleSuggestion";
-import { beccaErrorHandler } from "../../utils/beccaErrorHandler";
+import { SlashInt } from "../interfaces/slash/SlashInt";
+import { errorEmbedGenerator } from "../modules/commands/errorEmbedGenerator";
+import { handleResetLevels } from "../modules/commands/subcommands/manage/handleResetLevels";
+import { handleResetStars } from "../modules/commands/subcommands/manage/handleResetStars";
+import { handleSuggestion } from "../modules/commands/subcommands/manage/handleSuggestion";
+import { beccaErrorHandler } from "../utils/beccaErrorHandler";
 
 export const manage: SlashInt = {
   data: new SlashCommandBuilder()
