+++ src/commands/misc.ts
@@ -2,13 +2,13 @@ import {
   SlashCommandBuilder,
   SlashCommandSubcommandBuilder,
 } from "@discordjs/builders";
-import { SlashInt } from "../../interfaces/slash/SlashInt";
-import { errorEmbedGenerator } from "../../modules/commands/errorEmbedGenerator";
-import { handlePermissions } from "../../modules/commands/subcommands/misc/handlePermissions";
-import { handleSpace } from "../../modules/commands/subcommands/misc/handleSpace";
-import { handleUsername } from "../../modules/commands/subcommands/misc/handleUsername";
-import { handleXkcd } from "../../modules/commands/subcommands/misc/handleXkcd";
-import { beccaErrorHandler } from "../../utils/beccaErrorHandler";
+import { SlashInt } from "../interfaces/slash/SlashInt";
+import { errorEmbedGenerator } from "../modules/commands/errorEmbedGenerator";
+import { handlePermissions } from "../modules/commands/subcommands/misc/handlePermissions";
+import { handleSpace } from "../modules/commands/subcommands/misc/handleSpace";
+import { handleUsername } from "../modules/commands/subcommands/misc/handleUsername";
+import { handleXkcd } from "../modules/commands/subcommands/misc/handleXkcd";
+import { beccaErrorHandler } from "../utils/beccaErrorHandler";
 
 export const misc: SlashInt = {
   data: new SlashCommandBuilder()
